# Conflicts:
#	client/engine/core/Mob.cpp
#	client/engine/core/main.cpp

+++ client/engine/core/CommandSystem.cpp
@@ -56,7 +56,7 @@ void        CommandSystem::invalidateCommandAtTimestamp(timestamp time)
     }
 }
 
-int    CommandSystem::getSize() const
+unsigned long    CommandSystem::getSize() const
 {
     return _commands.size();
 }
